+++ src/app/(subpages)/blog/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from 'next';
 import { Locales } from '~/@types';
-import { Post } from '~/components/app/blog';
-import { getPosts } from '~/utils/contentful';
+import { BlogPostPreview } from '~/components';
+import { getPosts } from '~/utils';
 
 export const dynamic = 'force-static';
 
@@ -23,7 +23,7 @@ export default async function Page() {
   return (
     <div className="flex-grow self-center p-4 flex flex-col gap-8 w-full max-w-3xl">
       {posts.map((post) => {
-        return <Post key={post.slug} data={post} />;
+        return <BlogPostPreview key={post.slug} data={post} />;
       })}
     </div>
   );
