+++ src/components/app/blog/blog-post-preview.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import Link from 'next/link';
 import { PostType } from '~/@types';
 
-function PostPreview({ data }: { data: PostType }) {
+function BlogPostPreview({ data }: { data: PostType }) {
   return (
     <Link href={`/blog/${data.slug}`} className="hover:underline">
       <div className="grid grid-cols-[0_1fr] sm:grid-cols-[100px_1fr] grid-rows-2 bg-darkmode_regular rounded-md gap-4">
@@ -20,4 +20,4 @@ function PostPreview({ data }: { data: PostType }) {
   );
 }
 
-export { PostPreview };
+export { BlogPostPreview };
