+++ src/components/app/index.ts
@@ -1,0 +1,2 @@
+export * from './blog';
+export * from './home';
