+++ src/components/icons/clock-icon.component.tsx
@@ -1,6 +1,8 @@
-function ClockIcon() {
+import { SVGProps } from 'react';
+
+function ClockIcon(props?: SVGProps<SVGSVGElement>) {
   return (
-    <svg viewBox="0 0 24 24">
+    <svg viewBox="0 0 24 24" {...props}>
       <path d="M12 6v6h4.5m4.5 0a9 9 0 11-18 0 9 9 0 0118 0z" />
     </svg>
   );
