+++ src/components/icons/social/globe-icon.component.tsx
@@ -1,7 +1,9 @@
+import { SVGProps } from 'react';
+
 /**
  * Figma [link](https://www.figma.com/file/JOHGaqs67K52ZSj32xF8Bd/Weboldal?node-id=11%3A44) to component
  */
-function GlobeIcon(props?: React.SVGProps<SVGSVGElement>) {
+function GlobeIcon(props?: SVGProps<SVGSVGElement>) {
   return (
     <svg className="fill-white" viewBox="0 0 24 24" {...props}>
       <path d="M21.1714 19.7384C21.1768 19.7304 21.1848 19.7223 21.1902 19.7143C22.9446 17.6277 24 14.9384 24 12C24 9.06161 22.9446 6.37232 21.1929 4.28571C21.1875 4.27768 21.1795 4.27232 21.1741 4.26428C21.1446 4.22946 21.1179 4.19732 21.0884 4.16518C21.0777 4.15178 21.067 4.14107 21.0562 4.12768L20.9464 4.00178L20.9438 3.99911C20.9036 3.95357 20.8607 3.90803 20.8205 3.8625L20.8179 3.85982C20.7321 3.76875 20.6464 3.67768 20.558 3.58928L20.5554 3.58661L20.4268 3.45803L20.4188 3.45C20.3786 3.40982 20.3384 3.37232 20.2982 3.33482C20.2848 3.32143 20.2714 3.30803 20.2554 3.29464C20.2286 3.26786 20.2018 3.24375 20.175 3.21964C20.167 3.21161 20.1562 3.20357 20.1482 3.19286C18.0107 1.21071 15.1473 -9.53674e-07 12 -9.53674e-07C8.85268 -9.53674e-07 5.98929 1.21071 3.84911 3.19286C3.84107 3.20089 3.83036 3.20893 3.82232 3.21964C3.79554 3.24375 3.76875 3.27053 3.74196 3.29732C3.72857 3.31071 3.71518 3.32411 3.69911 3.3375C3.65893 3.375 3.61875 3.41518 3.57857 3.45268L3.57054 3.46071L3.44196 3.58928L3.43929 3.59196C3.35089 3.68036 3.26518 3.77143 3.17946 3.8625L3.17679 3.86518C3.13393 3.91071 3.09375 3.95625 3.05357 4.00178L3.05089 4.00446C3.01339 4.04464 2.97589 4.0875 2.94107 4.13036C2.93036 4.14375 2.91964 4.15446 2.90893 4.16786C2.87946 4.2 2.85268 4.23482 2.82321 4.26696C2.81786 4.275 2.80982 4.28036 2.80446 4.28839C1.05536 6.37232 0 9.06161 0 12C0 14.9384 1.05536 17.6277 2.80714 19.7143C2.8125 19.7223 2.82054 19.7304 2.82589 19.7384L2.90893 19.8375C2.91964 19.8509 2.93036 19.8616 2.94107 19.875L3.05089 20.0009C3.05089 20.0036 3.05357 20.0036 3.05357 20.0062C3.09375 20.0518 3.13393 20.0973 3.17679 20.1402L3.17946 20.1429C3.26518 20.2339 3.35089 20.325 3.43661 20.4134L3.43929 20.4161C3.48214 20.4589 3.52232 20.5018 3.56518 20.542L3.57321 20.55C3.66161 20.6384 3.75268 20.7241 3.84375 20.8071C5.98929 22.7893 8.85268 24 12 24C15.1473 24 18.0107 22.7893 20.1509 20.8071C20.2421 20.7235 20.3315 20.6378 20.4188 20.55L20.4268 20.542C20.4696 20.4991 20.5125 20.4589 20.5527 20.4161L20.5554 20.4134C20.6438 20.325 20.7295 20.2339 20.8125 20.1429L20.8152 20.1402C20.8554 20.0946 20.8982 20.0518 20.9384 20.0062C20.9384 20.0036 20.9411 20.0036 20.9411 20.0009C20.9786 19.9607 21.0161 19.9179 21.0509 19.875C21.0616 19.8616 21.0723 19.8509 21.083 19.8375C21.1133 19.8052 21.1428 19.7722 21.1714 19.7384ZM21.2812 15.9187C20.9116 16.792 20.4241 17.6009 19.8295 18.3348C19.1598 17.756 18.4276 17.2538 17.6464 16.8375C17.9571 15.5812 18.15 14.2018 18.2009 12.75H22.0446C21.9643 13.8455 21.7071 14.9089 21.2812 15.9187ZM22.0446 11.25H18.2009C18.15 9.79821 17.9571 8.41875 17.6464 7.1625C18.4312 6.74464 19.1625 6.24107 19.8295 5.66518C21.1221 7.25618 21.8953 9.20555 22.0446 11.25ZM15.9187 2.71875C16.9821 3.16875 17.9491 3.79018 18.8009 4.57232C18.3061 4.99364 17.7735 5.36844 17.2098 5.69196C16.7893 4.48661 16.2509 3.43929 15.6241 2.60089C15.7232 2.63839 15.8223 2.67857 15.9187 2.71875ZM13.492 21.4848C13.2455 21.6777 12.9991 21.825 12.75 21.9241V16.9554C13.8128 17.0295 14.858 17.2662 15.8491 17.6571C15.6268 18.3161 15.3696 18.9241 15.0723 19.4732C14.6063 20.3411 14.0598 21.0348 13.492 21.4848ZM15.0723 4.52678C15.367 5.07857 15.6268 5.68661 15.8491 6.34286C14.858 6.73381 13.8128 6.97049 12.75 7.04464V2.07857C12.9964 2.17768 13.2455 2.32232 13.492 2.51786C14.0598 2.96518 14.6063 3.65893 15.0723 4.52678ZM12.75 15.4527V12.75H16.7009C16.658 13.9339 16.5107 15.083 16.2643 16.1732L16.2563 16.2054C15.1321 15.7791 13.95 15.5253 12.75 15.4527ZM12.75 11.25V8.54732C13.9768 8.47232 15.1554 8.2125 16.2563 7.79464L16.2643 7.82679C16.5107 8.91696 16.658 10.0634 16.7009 11.25H12.75ZM11.25 12.75V15.4527C10.0232 15.5277 8.84464 15.7875 7.74375 16.2054L7.73571 16.1732C7.48929 15.083 7.34196 13.9366 7.29911 12.75H11.25ZM7.29911 11.25C7.34196 10.0661 7.48929 8.91696 7.73571 7.82679L7.74375 7.79464C8.84464 8.2125 10.0205 8.47232 11.25 8.54732V11.25H7.29911ZM11.25 16.9554V21.9214C11.0036 21.8223 10.7545 21.6777 10.508 21.4821C9.94018 21.0348 9.39107 20.3384 8.925 19.4705C8.63036 18.9187 8.37054 18.3107 8.14821 17.6545C9.14464 17.2634 10.1812 17.0304 11.25 16.9554ZM11.25 7.04464C10.1872 6.97049 9.14199 6.73381 8.15089 6.34286C8.37321 5.68393 8.63036 5.07589 8.92768 4.52678C9.39375 3.65893 9.94018 2.9625 10.5107 2.51518C10.7571 2.32232 11.0036 2.175 11.2527 2.07589V7.04464H11.25ZM8.08125 2.71875C8.18036 2.67857 8.27679 2.63839 8.37589 2.60089C7.74911 3.43929 7.21071 4.48661 6.79018 5.69196C6.22768 5.37053 5.69464 4.99553 5.19911 4.57232C6.05089 3.79018 7.01786 3.16875 8.08125 2.71875ZM2.71875 8.08125C3.08839 7.20803 3.57589 6.39911 4.17054 5.66518C4.8375 6.24107 5.56875 6.74464 6.35357 7.1625C6.04286 8.41875 5.85 9.79821 5.79911 11.25H1.95536C2.03571 10.1545 2.29286 9.09107 2.71875 8.08125ZM1.95536 12.75H5.79911C5.85 14.2018 6.04286 15.5812 6.35357 16.8375C5.57245 17.2538 4.84024 17.756 4.17054 18.3348C2.87794 16.7438 2.10473 14.7945 1.95536 12.75ZM8.08125 21.2812C7.01786 20.8313 6.05089 20.2098 5.19911 19.4277C5.69464 19.0045 6.22768 18.6321 6.79018 18.308C7.21071 19.5134 7.74911 20.5607 8.37589 21.3991C8.27679 21.3616 8.17768 21.3214 8.08125 21.2812ZM15.9187 21.2812C15.8196 21.3214 15.7232 21.3616 15.6241 21.3991C16.2509 20.5607 16.7893 19.5134 17.2098 18.308C17.7723 18.6295 18.3054 19.0045 18.8009 19.4277C17.9538 20.2066 16.9788 20.8336 15.9187 21.2812Z" />
