+++ src/components/icons/social/instagram-icon.component.tsx
@@ -1,7 +1,9 @@
+import { SVGProps } from 'react';
+
 /**
  * Figma [link](https://www.figma.com/file/JOHGaqs67K52ZSj32xF8Bd/Weboldal?node-id=11%3A44) to component
  */
-function InstagramIcon(props?: React.SVGProps<SVGSVGElement>) {
+function InstagramIcon(props?: SVGProps<SVGSVGElement>) {
   return (
     <svg viewBox="0 0 24 24" {...props}>
       <path d="M11.9968 7.99815C9.79333 7.99815 7.99515 9.79633 7.99515 11.9998C7.99515 14.2032 9.79333 16.0014 11.9968 16.0014C14.2002 16.0014 15.9984 14.2032 15.9984 11.9998C15.9984 9.79633 14.2002 7.99815 11.9968 7.99815ZM23.9987 11.9998C23.9987 10.3427 24.0137 8.70061 23.9206 7.04652C23.8275 5.12526 23.3893 3.42014 21.9843 2.01522C20.5764 0.607291 18.8743 0.172005 16.953 0.0789442C15.2959 -0.014117 13.6539 0.000892919 11.9998 0.000892919C10.3427 0.000892919 8.70061 -0.014117 7.04652 0.0789442C5.12526 0.172005 3.42014 0.610293 2.01522 2.01522C0.607291 3.42314 0.172005 5.12526 0.0789442 7.04652C-0.014117 8.70361 0.000892919 10.3457 0.000892919 11.9998C0.000892919 13.6539 -0.014117 15.2989 0.0789442 16.953C0.172005 18.8743 0.610293 20.5794 2.01522 21.9843C3.42314 23.3923 5.12526 23.8275 7.04652 23.9206C8.70361 24.0137 10.3457 23.9987 11.9998 23.9987C13.6569 23.9987 15.2989 24.0137 16.953 23.9206C18.8743 23.8275 20.5794 23.3893 21.9843 21.9843C23.3923 20.5764 23.8275 18.8743 23.9206 16.953C24.0167 15.2989 23.9987 13.6569 23.9987 11.9998ZM11.9968 18.1568C8.58954 18.1568 5.83973 15.407 5.83973 11.9998C5.83973 8.59254 8.58954 5.84273 11.9968 5.84273C15.404 5.84273 18.1538 8.59254 18.1538 11.9998C18.1538 15.407 15.404 18.1568 11.9968 18.1568ZM18.406 7.02851C17.6105 7.02851 16.968 6.38609 16.968 5.59057C16.968 4.79504 17.6105 4.15262 18.406 4.15262C19.2015 4.15262 19.8439 4.79504 19.8439 5.59057C19.8442 5.77947 19.8071 5.96656 19.735 6.14112C19.6628 6.31569 19.5569 6.4743 19.4233 6.60787C19.2897 6.74145 19.1311 6.84736 18.9565 6.91954C18.782 6.99172 18.5949 7.02875 18.406 7.02851Z" />
