+++ src/components/icons/social/phone-icon.component.tsx
@@ -1,7 +1,9 @@
+import { SVGProps } from 'react';
+
 /**
  * Figma [link](https://www.figma.com/file/JOHGaqs67K52ZSj32xF8Bd/Weboldal?node-id=11%3A44) to component
  */
-function PhoneIcon(props?: React.SVGProps<SVGSVGElement>) {
+function PhoneIcon(props?: SVGProps<SVGSVGElement>) {
   return (
     <svg viewBox="0 0 24 24" {...props}>
       <path d="M20 15.5C18.75 15.5 17.55 15.3 16.43 14.93C16.08 14.82 15.69 14.9 15.41 15.17L13.21 17.37C10.38 15.93 8.06 13.62 6.62 10.79L8.82 8.58C9.1 8.31 9.18 7.92 9.07 7.57C8.7 6.45 8.5 5.25 8.5 4C8.5 3.45 8.05 3 7.5 3H4C3.45 3 3 3.45 3 4C3 13.39 10.61 21 20 21C20.55 21 21 20.55 21 20V16.5C21 15.95 20.55 15.5 20 15.5ZM12 3V13L15 10H21V3H12Z" />
