+++ src/components/index.ts
@@ -7,4 +7,5 @@ export * from './profile';
 export * from './seo';
 export * from './student-group';
 export * from './lightbox';
-export * from './loading';
+export * from './spinner';
+export * from './app';
