+++ src/components/spinner/spinner.component.tsx
