+++ src/components/student-group/student-group.component.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from 'react';
 import Image from 'next/image';
 import type { StudentGroupType } from '~/@types';
-import { ChevronIcon, SocialIcon } from '../icons';
+import { ChevronIcon, SocialIcon } from '..';
 
 function StudentGroup({ name, logo, description, socials = [], isDense = false }: StudentGroupType & { isDense?: boolean }) {
   const [isOpen, setIsOpen] = useState(false);
