+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './functions';
 export * from './contentful-renderer';
 export * from './defaults';
+export * from './contentful';
